Separated out private checksum operations
